+++ packages/blaze-emulator/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @blaze-cardano/emulator
 
+## 0.1.32
+
+### Patch Changes
+
+- a28ff6a: feat: Stake registration balancing
+- Updated dependencies [a28ff6a]
+  - @blaze-cardano/tx@0.0.27
+  - @blaze-cardano/wallet@0.1.12
+
+## 0.1.31
+
+### Patch Changes
+
+- Updated dependencies [99ce4a5]
+  - @blaze-cardano/tx@0.0.26
+  - @blaze-cardano/wallet@0.1.11
+
 ## 0.1.30
 
 ### Patch Changes