+++ packages/blaze-tx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @blaze-cardano/tx
 
+## 0.0.27
+
+### Patch Changes
+
+- a28ff6a: feat: Stake registration balancing
+
+## 0.0.26
+
+### Patch Changes
+
+- 99ce4a5: feat: addRegisterStake
+
 ## 0.0.25
 
 ### Patch Changes