+++ packages/blaze-uplc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @blaze-cardano/uplc
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [a28ff6a]
+  - @blaze-cardano/tx@0.0.27
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [99ce4a5]
+  - @blaze-cardano/tx@0.0.26
+
 ## 0.0.7
 
 ### Patch Changes