+++ packages/blaze-wallet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @blaze-cardano/wallet
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [a28ff6a]
+  - @blaze-cardano/tx@0.0.27
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [99ce4a5]
+  - @blaze-cardano/tx@0.0.26
+
 ## 0.1.10
 
 ### Patch Changes